Moon Dev Testing
Several different moons created to test deforming it with an explosion. Best so far is the "Moon_Multi-Obj". The "explosion" object in the scene will delete the minimoons upon collision. The explosion will be instatiated at the point where the rocket hits the moon. The explosion will be a cylinder with the radius determined by the blast power. The explosions height is calculated based on the Moon's Diameter.  The Moon calculates health based on number of minimoons created.